+++ vk_sdk/__init__.py
@@ -3,10 +3,13 @@
 :authors: SPRAVEDLIVO
 :license: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, see LICENSE file
 
-:copyright: (c) 2021 SPRAVEDLIVO
+:copyright: (c) 2022 SPRAVEDLIVO
 """
-
+from .lp import BotLongPoll, UserLongPoll
+from .user import User
+from .keyboard import Keyboard
+from .cmd import start_command, command, after_text_matcher, after_func
 
 __author__ = 'SPRAVEDLIVO'
-__version__ = '1.5.3'
+__version__ = '1.6.1'
 __email__ = 'admin@spravedlivo.dev'