+++ vk_sdk/imports.py
@@ -1,6 +1,7 @@
 import os
 
 from importlib import util
+import importlib
 
 
 class ImportTools(object):
@@ -22,8 +23,28 @@ class ImportTools(object):
                 thisPath = os.path.join(path, file)
                 if os.path.isdir(thisPath):
                     continue
-                fileName = os.path.splitext(file)[0]
-                spec = util.spec_from_file_location(fileName, thisPath)
-                foo = util.module_from_spec(spec)
-                spec.loader.exec_module(foo)
-                self.modules.update({fileName: foo})
+                self.imp_by_path(thisPath)
+                self.imported.append(path)
+
+    def reload(self, module):
+        for k, v in self.modules.items():
+            if k == module:
+                self.modules[k] = importlib.reload(v)
+
+    def reload_all(self):
+        for k, v in self.modules.items():
+            self.modules[k] = importlib.reload(v)
+
+    @classmethod
+    def imp_by_path(cls, path):
+        if not path.endswith(".py"):
+            path += ".py"
+        module_name = os.path.splitext(os.path.basename(path))[0]
+        spec = util.spec_from_file_location(module_name, path)
+        foo = util.module_from_spec(spec)
+        spec.loader.exec_module(foo)
+        cls.modules[path.replace(os.path.sep, "/")] = foo
+        return foo
+
+
+require = ImportTools.imp_by_path