+++ vk_sdk/thread.py
@@ -96,7 +96,7 @@ def every(interval, *myArgs, callback=None, **myKwargs):
 
 main_thread = None
 
-@events.once("pre_start")
+@events.once("start")
 def update_main_thread():
     global main_thread
     main_thread = ThreadManager.threadByName("Main")